requirements.txt
# Core API
fastapi==0.115.2
uvicorn==0.30.6
pydantic==2.8.2
python-multipart==0.0.9
Pillow==10.3.0

# AWS SDK
boto3==1.35.2
botocore==1.35.2

# Transformers & BLIP2 Model
transformers==4.44.2
torch==2.4.0
optimum==1.20.0

# Data / Utilities
numpy==1.26.4
python-dotenv==1.0.1
httpx==0.27.0

# Optional Local ML Runtime
accelerate==0.33.0
sentencepiece==0.2.0

# For ECS container runtime and health
requests==2.32.3

# Logging & Observability
loguru==0.7.2

# Test / Validation
pytest==8.3.1
